Better handling of separator directives and feature selection (#75)
This update mashes together two changes. First, it gives finer control to how and when separator directives are printed after complex features. Second, it enables selection of multiple feature types simultaneously in the `tag.select.features` function and updates the `bae` module to correctly use this function in `traverse` mode.